# Remove all uses of *[]*Foo