--- a/oauth_manager_test.go
+++ b/oauth_manager_test.go
@@ -103,7 +103,7 @@ func getOAuthChain(spec *APISpec, Muxer *mux.Router) {
 func makeOAuthAPI(t *testing.T) *APISpec {
 	spec := createSpecTest(t, oauthDefinition)
 
-	specs := &[]*APISpec{spec}
+	specs := []*APISpec{spec}
 	newMuxes := mux.NewRouter()
 	loadAPIEndpoints(newMuxes)
 	loadApps(specs, newMuxes)