This only makes sense if the variable is a parameter and it is modified,
e.g. appended to. None of these fall under that case.

Note that there are still some `*[]Foo` around, but I'm leaving those for another day.

@lonelycode PTAL