@buger 2017-02-01
I wonder why it was even used originally. Arrays are already pointers, same as maps.