@mvdan 2017-02-01
> Arrays are already pointers, same as maps.

Careful with the wording there - you mean slices, not arrays :)

Also, grouping slices with maps isn't exactly correct. Maps are a single pointer, while slices are a three-field struct (arr, len, cap) where the actual backing array is a pointer. Using pointers for slices is still useful to update any of the three fields, as explained above.