@buger 2017-02-01
Well, for sure I meant slices.

But when they used as arguments, they send as pointers (yes to struct), not by value. So this is still technically a pointer. And you can access real pointer to "slice" anytime, by dereferencing it anyway, without explicitly sending(or returning it) to function.

Anyway, I did not saw usage of such functionality anywhere, that's why I'm asking, maybe we missing something.