@buger 2017-02-01
Actually, I may be wrong about dereferencing it, so it will be possible to update capacity. Still, we not used it anyway :)