@lonelycode 2017-02-02
The reason it's this way is to stop copying API Definitions around. Using this was just lighter and faster. We would need to confirm whether this change has a detrimental effect on performance and load times of large API Definitions sets for a gateway (2,000 and up). Until we can be sure of that am reticent to make a change.