@mvdan 2017-02-02
Slices don't actually contain the objects directly, so I doubt this would affect performance in any way. Rather the opposite - you're using pointers to 24-byte structs, which is likely to degrade performane a tiny bit.

But my main concern here is making the code overly obtuse for no reason.

This isn't blocking anything else of mine though, so feel free to run your tests first.