@lonelycode 2017-02-02
> so I doubt this would affect performance in any way. Rather the opposite - you're using pointers to 24-byte structs, which is likely to degrade performane a tiny bit.

Assumption is the mother of all f*ckups, prove that performance isn't affected and I'm a happy man.