@mvdan 2017-02-02
Well, I would if we had benchmarks :) I don't know how you run yours, so I can't prove much.