@buger 2017-02-02
Well, while I agree that performance should be ok, try very basic synthetic benchmark of inserting few thousand APISpecs and accessing few fields in contexts of *[] and [], it should be enough.