@mvdan 2017-02-02
I've opened https://github.com/TykTechnologies/tyk/issues/453 about our gateway benchmark discussion on Slack.

If I write an APISpec benchmark, would it be useful to keep in the repo?