@buger 2017-02-02
This synthetic test unlikely will be really useful, so probably you can just show results and throw it away. But in the context of #453 , it can be useful to add some benchmarks which test specific parts, but it will be separate task anyway.