@lonelycode 2017-02-02
> Well, while I agree that performance should be ok, try very basic synthetic benchmark of inserting few thousand APISpecs, iterating over them, and accessing few fields in contexts of *[] and [], it should be enough.

The synthetic test should be:

- Insert a few thousand APIs
- Hot reload them a few times
- Check speed and memory consumption