@mvdan 2017-02-09
I've added a somewhat stupid benchmark:

```
name               old time/op    new time/op    delta
ApiInsertReload-4    14.9ms ± 1%    14.8ms ± 0%  -0.48%  (p=0.015 n=6+6)

name               old alloc/op   new alloc/op   delta
ApiInsertReload-4    4.37MB ± 0%    4.37MB ± 0%    ~     (p=0.485 n=6+6)

name               old allocs/op  new allocs/op  delta
ApiInsertReload-4     47.2k ± 0%     47.2k ± 0%    ~     (p=0.608 n=6+6)
```

I say it's stupid because it doesn't use http to create the apis, and it doesn't reload them.

The way the tests are written is a bit confusing:

* do I need to set up `http.DefaultServeMux`?
* do I need to use `httptest`?
* don't I need a secret to do these things, like shown in https://tyk.io/tyk-documentation/get-started/with-tyk-on-premise/tutorials/tyk-premise-community-edition/create-an-api/?

Also, @lonelycode, when you said "Insert a few thousand APIs", do you mean one by one or all at once?