@mvdan 2017-02-09
Actually, ignore the numbers above as the gateway is erroring under the covers and not actually adding the endpoints.