@mvdan 2017-02-09
Actually, these are the real numbers for the silly benchmark - I hadn't noticed the gateway was erroring and not actually doing much.

```
name               old time/op    new time/op    delta
ApiInsertReload-4    77.7ms ± 0%    76.9ms ± 1%  -1.07%  (p=0.030 n=5+6)

name               old alloc/op   new alloc/op   delta
ApiInsertReload-4    33.7MB ± 0%    33.7MB ± 0%    ~     (p=0.937 n=6+6)

name               old allocs/op  new allocs/op  delta
ApiInsertReload-4      400k ± 0%      400k ± 0%    ~     (p=0.619 n=6+6)
```