@mvdan 2017-02-10
Okay - I'll open an issue about making the benchmark better.